comment

+++ src/main/java/life/guohui/community/service/QuestionService.java
@@ -49,7 +49,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public PaginationDTO listByUserId(Integer userId, Integer page, Integer size) {
+    public PaginationDTO listByUserId(Long userId, Integer page, Integer size) {
         PaginationDTO paginationDTO = new PaginationDTO();
         Integer totalPage;
         //拿到总数
@@ -77,7 +77,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public QuestionDTO getById(Integer id) {
+    public QuestionDTO getById(Long id) {
         Question question = questionMapper.selectByPrimaryKey(id);
         if(question == null){
             throw new CustomizeException(CustomizeErrorCode.QUESTION_NOT_FOUND);
@@ -114,7 +114,7 @@ public class QuestionService {
 
     }
 
-    public void inView(Integer id) {
+    public void inView(Long id) {
         Question question = new Question();
         question.setId(id);
         question.setViewCount(1);
